pyproject.toml
[project]
name = "versaai"
version = "0.2.0"
description = "VersaAI: Local AI inference server with OpenAI-compatible API"
authors = [
    {name = "The No-hands Company", email = "info@nohands.company"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Proprietary - See EULA.txt"}

dependencies = [
    # API Server (core — always needed)
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # HTTP / inference provider clients
    "httpx>=0.27.0",

    # Config
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",

    # CLI
    "rich>=13.0.0",
    "click>=8.1.0",

    # Logging
    "structlog>=23.0.0",
]

[project.optional-dependencies]
# Heavy ML deps — opt-in only
ml = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "accelerate>=0.24.0",
    "llama-cpp-python>=0.2.0",
    "sentence-transformers>=2.2.0",
]

# RAG pipeline
rag = [
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-core>=0.1.0",
    "faiss-cpu>=1.7.0",
    "beautifulsoup4>=4.12.0",
]

# All optional deps
all = [
    "versaai[ml,rag]",
]

dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",

    # Code quality
    "ruff>=0.5.0",
    "mypy>=1.7.0",

    # Build tools
    "pybind11>=2.11.0",
    "cmake>=3.25.0",
    "ninja>=1.11.0",
]

[project.scripts]
versaai = "versaai.cli:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["versaai*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"

select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]

ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[[tool.mypy.overrides]]
module = [
    "transformers.*",
    "langchain.*",
    "chromadb.*",
    "llama_cpp.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=versaai",
    "--cov-report=term-missing",
]

[tool.coverage.run]
source = ["versaai"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
